Move to src and format code.
