+++ Sming/Components/MultipartParser/component.mk
@@ -1,2 +1,5 @@
+COMPONENT_SRCDIRS := src
+COMPONENT_INCDIRS := src
+
 COMPONENT_SUBMODULES	:= multipart-parser
 COMPONENT_SRCFILES		:= multipart-parser/multipart_parser.c
